完成部分接口

+++ app/service/v1/system/department.js
@@ -11,15 +11,35 @@ class UsersService extends BaseService {
     this.modelName = 'Departments'
   }
 
-  // 删除
-  async destroy (ids) {
-    return await this.ctx.model[this.modelName].destroy({
-      where: {
-        deptId: {
-          [Op.or]: ids
-        }
+  // 刪除id
+  async destroy (deptId) {
+    try {
+      // 建立事务对象
+      let transaction = await this.ctx.model.transaction();
+      
+      // 事务增操作
+      let menus = await this.ctx.model[this.modelName].findAll({
+        where: {
+          parentId: deptId
+        },
+        transaction
+      });
+      if (menus.length) {
+        throw Error('含有子元素，不能删除')
       }
-    })
+
+      // 刪除数据
+      await this.ctx.model[this.modelName].destroy({
+        where: {
+          deptId
+        }
+      })
+      // 提交事务
+      await transaction.commit();
+      return true
+    } catch (error) {
+      this.ctx.throw(500, error)
+    }
   }
 }
